Parser refactor:
Introduce an implementation-only struct (`Parser`) which holds to token iterator.
This means parse functions don't have to constantly pass around the token iterator, it is stored as a field.
Better helpers to construct AST nodes, and unique pointers to AST nodes
Better helpers to construct a span of tokens when making ast nodes
Operator precedence is a template parameter.