DotVisitor refactor:
Actually use a tree structure when creating dot graph, instead of the weird state machine that was there before.
This tree structure is infinitely easier to reason about, with the only downside being memory use from the DotNode structs, but these may even be optimized out by the compiler anyway.
Additionally, since this version writes to the output stream only in the destructor, it will not produce a partial result upon program crash. This *is* a downside, but a program crash shouldn't really be happening anyway.
The writing could be made marginally earlier, with only benefits in the "untyped" case; see the TODO in dot_visitor.cpp