Run struct fields through type walker
This means struct fields now have a strong type bound to them and `convert_type` does not have to be abused as often.
Additionally, `convert_type` itself was marked as deprecated (simply to find references to it).
This now-deprecated method will probably move from the Compiler directly to the TypeWalker, which is the only class which has a justifiable use for the method in the first place.