COMPLETELY rewrite overload selection and implicit cast algorithm
The whole system is split up across multiple files and classes. Compared to before, overload selection is done in multiple steps:
First determine if an argument can be implicitly converted to a parameter (and keep track of what was done)
Then, compare all overloads to find the best one. Overloads are rated by comparing their conversions each parameter at a time
Instead of giving a total rating to the function, each argument is compared in lockstep

Since implicit conversion steps are kept track of, the actual conversion in the compilation step is much simpler

Documentation for the new overload-related stuff isn't fully written yet