Language change: Slice literals now require a colon between the type name and the opening [
This is to distinguish it better from slice types, especially in the case of the empty slice, which was previously sorta ambiguous.

For example, previously, `I32[]` would have been an empty slice of I32, but it is also the type `I32[]`. Since a type on its own isn't a valid expression, there was no ambiguity, but this may be wanted in the future.

Now, an empty slice literal is `I32:[]`.