Move end of scope check responsibility to compiler
I will probably regret this later, but it is currently very hard to insert arbitrary ast nodes before or after existing ones in the type walker.
i.e., when leaving a scope with a return, I want to insert DtorExprs in *front* of the return.
This is more of a limitation of the ast system and will definitely cause technical debt