Implement lambdas with captures (closures)
To perform this, every function object is actually a struct with a pointer to a function, and a pointer to a type-erased "closure object", which is passed to the lambda. The lambda then unpacks it accordingly

Note that there is no safety yet, such that captured values do not outlive the values they reference.

Note that all values are currently captured exactly as the type they were (i.e. an `I32 mut` is captured as exactly that, even if it is never modified inside of the lambda)