Refactor most logic related to generics
A lot of changes at once, but the gist is to be able to support non-type
generic parameters in the future. A type with a non-type generic
parameter (StaticArray) has been added to std.ym, but it is never
instantiated and doesn't fully work yet.

Also improves how generics are instantiated by a lot, instead of
polluting everything with duplicate instantiations for each method in
a generic struct, those methods themselves are made to be generic
instead.

It won't be perfect, but it's at least closer.

Given the large amount of changes, there will 100% be something broken.

Also, there is lots of ugly repetitive and deprecated methods, which
should be cleaned up in a future commit.