Further simplify Substitutions logic
Removes the parent pointer, which added needless complexity. Now, all
instances of Substitutions own all of their values, which was the case
before anyways, due to defensive deep copies, but is now automatic
thanks to value semantics. This also solves the issue of comparisons
being expensive due to, essentially, recursion.